[app] Create new DVR-Scan application for GUI
