+++ .gitignore
@@ -2,6 +2,7 @@
 *.pyc
 *.pyo
 *.egg-info/
+__pycache__/
 
 # Location of built documentation
 site/
