+++ setup.cfg
@@ -44,9 +44,12 @@ install_requires =
     screeninfo
     tqdm
 packages =
+# Main application
     dvr_scan
     dvr_scan.app
     dvr_scan.cli
+    dvr_scan.shared
+# Documentation (pre-built HTML but without images to keep size small)
     dvr_scan.docs
     dvr_scan.docs.assets
     dvr_scan.docs.assets.javascripts
@@ -65,6 +68,7 @@ opencv-headless = opencv-python-headless
 [options.entry_points]
 console_scripts =
     dvr-scan = dvr_scan.__main__:main
+    dvr-scan-app = dvr_scan.app.__main__:main
 
 [aliases]
 test = pytest
